+++ src/coords/angle.js
@@ -5,6 +5,14 @@
 import { calibrate_hmsn } from '../time';
 import { NaiveTime } from '../chrono';
 
+const ROUND_DIGITS = 100_000;
+
+/**
+ * @param {number} num
+ */
+const round = num =>
+  Math.round(num * ROUND_DIGITS) / ROUND_DIGITS;
+
 /**
  * @typedef NaiveTimeContext
  * @type {import('../chrono/naive_time.js').NaiveTimeContext}
@@ -24,8 +32,34 @@ import { NaiveTime } from '../chrono';
  */
 
 /**
- * A context (or an instance)
- * to be created.
+ * This is a context (or an instance)
+ * returned from the factory.
+ *
+ * When you create the instance, you
+ * may freely choose to manage your
+ * 'hour' as in that of time, or in
+ * that of the degree context
+ * (e.g. '24 hours' vs '360 degrees').
+ * However, when you run 'calibrate'
+ * method, it defaults to oveflow
+ * the 'hour' in the time context.
+ * If you wish to calibrate 'hour'
+ * in the dgree context, then you
+ * need to explicitly specify
+ * 'calibrate' to run in the degree
+ * context by giving the option
+ * 'options.angle' to 'calibrate'.
+ *
+ * Moreover, it defaults NOT to
+ * calibrate 'hour' even when a negative
+ * value were specified to 'hour'.
+ * For instance, say, you had '-10' for
+ * your 'hour', it will stay '-10' even
+ * if you run 'calibrate'. If you want
+ * to change the negative into the
+ * positive, you need to explicitly
+ * specify 'options.hour_overflow'
+ * to 'calibrate' method.
  *
  * @typedef AngleContext
  * @type {Object}
@@ -64,6 +98,9 @@ import { NaiveTime } from '../chrono';
 
 /**
  * @callback Calibrate
+ * @param {Object} options
+ * @param {boolean} [options.angle=false] - When specified, it will take the given data as angle-based, and will have 'hour' overflow only when it reached 360. Usually, it should be 24 as a default, and this option should be set to 'false'. When set to 'true', it will set the said limit to 360.
+ * @param {boolean} [options.hour_overflow=true] - If specified TRUE, changes the negative into the positive. If FALSE, let the negative hour as it is.
  * @returns {number} - Excess days resulting due to the calibration.
  * @see {@link module:sowngwala/time.calibrate_hmsn}
  */
@@ -78,20 +115,9 @@ import { NaiveTime } from '../chrono';
  */
 export const Angle = Object.freeze({
   from_hms,
-  // from_degree_ms,
 });
 
 /**
- * Watch out when you have 'degree' instead
- * of 'hour', and you want an instance created.
- * Actually, don't. Instead, convert it to
- * decimal degrees, and then to Hour, Minute,
- * and Second for creating 'AngleContext'.
- *
- * TODO:
- * We may want another static method to create
- * from Degree-Minute-Second, maybe?
- *
  * @public
  * @static
  * @type {FromHMS}
@@ -140,7 +166,7 @@ function _from_hms(h, m, s) {
     day_excess: () => day_excess,
     to_naive_time,
     calibrate,
-    print: () => `${hour}°${min}'${sec}`,
+    print: () => `${hour}°${min}'${round(sec)}`,
   });
 
   /**
@@ -157,7 +183,21 @@ function _from_hms(h, m, s) {
    * @protected
    * @type {Calibrate}
    */
-  function calibrate() {
+  function calibrate(options) {
+    // Usually, take it in time context
+    // (defaults to FALSE).
+    // Only when specified TRUE,
+    // take it as a degree context.
+    const angle = !!options?.angle;
+
+    // Usually, let the negative hour
+    // stay negative as it is
+    // (defaults to FALSE).
+    // Only when specified TRUE,
+    // then change the negative
+    // into the positive.
+    const hour_overflow = !!options?.hour_overflow;
+
     ({
       hmsn: { hour, min, sec },
       day_excess,
@@ -166,6 +206,8 @@ function _from_hms(h, m, s) {
       min,
       sec,
       nano: 0,
+      angle,
+      hour_overflow,
     }));
 
     return day_excess;