+++ types/time/calibrate_hmsn.d.ts
@@ -1,4 +1,4 @@
-export function calibrate_hmsn({ hour, min, sec, nano, angle, }: CalibrateArguments): {
+export function calibrate_hmsn({ hour, min, sec, nano, angle, hour_overflow, }: CalibrateArguments): {
     hmsn: HMSNano;
     day_excess: DecimalDays;
 };
@@ -9,6 +9,7 @@ export type NanoSecond = import('../types.js').NanoSecond;
 export type DecimalDays = import('../types.js').DecimalDays;
 export type AdditionalOptions = {
     angle?: boolean;
+    hour_overflow?: boolean;
 };
 export type CalibrateArguments = HMSNano & AdditionalOptions;
 export type HMSNano = {